Various improvements around distance and pace/speed

+++ TriResultsV2/Helpers/SwimHelper.cs
@@ -7,9 +7,9 @@ namespace TriResultsV2.Helpers
 {
     public static class SwimHelper
     {
-        public static string Get25MetreSwimPace(int distanceInMetres, TimeSpan swimTime)
+        public static string Get25MetreSwimPace(double distanceInMetres, TimeSpan swimTime)
         {
-            int numberOf25MetreLaps = distanceInMetres / 25;
+            double numberOf25MetreLaps = distanceInMetres / 25;
             double secondsPer25MetreLap = swimTime.TotalSeconds / numberOf25MetreLaps;
             string swimPace = $"{Math.Round(secondsPer25MetreLap, 1)} sec/25m";
             return swimPace;
